Rename command and make it single
