+++ packages/openapi-generator/bin/run
@@ -1,5 +1,16 @@
 #!/usr/bin/env node
+/* eslint-disable */
 
-require('@oclif/command').run()
-.then(require('@oclif/command/flush'))
-.catch(require('@oclif/errors/handle'))
+const fs = require('fs');
+const path = require('path');
+const project = path.join(__dirname, '../tsconfig.json');
+const dev = fs.existsSync(project);
+
+if (dev) {
+  require('ts-node').register({ project });
+}
+
+const command = require(`../${dev ? 'src/cli' : 'dist/cli'}`);
+  command.run()
+  .catch(require('@oclif/errors/handle'));
+
