+++ packages/openapi-generator/src/generator.ts
@@ -5,7 +5,7 @@ import { resolve } from 'path';
 import { createLogger, errorWithCause } from '@sap-cloud-sdk/util';
 import execa = require('execa');
 import { OpenAPIV3 } from 'openapi-types';
-import { GeneratorOptions } from './commands/generate-rest-client';
+import { GeneratorOptions } from './options';
 import { apiFile, indexFile, createFile } from './wrapper-files';
 import { OpenApiDocument } from './openapi-types';
 import { parseOpenApiDocument } from './parser';
