+++ packages/openapi-generator/src/options.ts
@@ -1,0 +1,5 @@
+export interface GeneratorOptions {
+  inputDir: string;
+  outputDir: string;
+  clearOutputDir?: boolean;
+}
